add TempCodeAuthApiController
